Added stage locking and round robbin functionality for multiple servers.
This feature basically allows you to define a set of stages (names basically)
inside of the /etc/octopusci.yml file. Each job will consume one stage while
it is running and give it back when its is done running. This allows you to
have N different servers for running your tests and N different workers.

Note: Octopusci should eventually be responsible launching the correct number
of workers for you in the future.