Made the stages and stage locking an optional feature.
The way it determines if it should have the feature enabled/disabled is
by checking if the config file defined a list called 'stages' if the key
'stages' exists at the top level then it will enable stages, round robin them,
and enable locking for them. If 'stages' key is not found at the top level of
the config file then it will not use stages and therefore will not lock them
or round robin them.