Added Bundler.with_clean_env to run Job run method.
This is needed because when Octopusci is run which uses Bundler it leads a
number of environment variables up with values. This is a problem because if
the Job run method runs any commands that depend on environment variables or
the lack there of, such as `bundle install`, you will end with issues.
Luckily Bundler provides a top level method to help avoid this which basically
temporarily resets the env back to the state it was in before Octopusci
loaded Bundler. Then it runs the block passed to it and then it sets it back
to what Bundler wants in the env.