Reworked tentacles to just launch/manage workers in the background. gh-10
This corrects the issue that was happening before where tentacles would only
ever launch one registered worker. This was because Resque workers are
designed and expected to be independent forked and detached processes. Hence,
we can't have the parent process wait for them. Therefore, this commit
resolves this issue by making tentacles a management tool for starting,
stoping, and restarting the Resque workers similar to a normal init script
that would be found in /etc/init.d/ on a linux system. Note: It stores the
PIDs of the workers in redis so if you are using this in the system startup
it needs to happen after redis has already started up.