Switched to hash format instead of OpenStruct. Switched to string keys instead of symbols.
I switched to hash format because OpenStruct really wasn't what we needed, it was just a poor attempt at what I had hoped would be a quick fix. The string keys change was made simply so that all keys are consistent instead of having some be symbols and others be strings because it was causing confusion.