Bunch of cleanup and polish. Nuked no longer needed db based code.
I worked in a bunch of polish and nuked all the old code I was using for reference since we no longer talk to the database at all. I also update the gem spec to get rid of the mysql database dependency. Beyond that I updated the octopusci-skel app to generate a more up to date example config.