Added version output to the ocotpusci-reset-redis command (#28)
I made this change so that users could easily identify what version of Octopusci they are using when they run into bugs. I also modified the script to be safer by outputting a warning explaining the dangers of the script when run with no args/opts. Now, in order to get it to actually reset the redis data store you have to run the command with the -r switch.