Modified octopusci-tentacles to no longer reset the stage locker on every call (#21).
I made this modification because it was causing problems when restarting Octopusci after an upgrade and a few other situations. The problem was that we were getting in states where multiple jobs were running on a single stage because of the reset. So, now I have made it so octopusci-tentacles only resets the stage locker based on the config stages if it doesn't already exist. If it already exists then octopusci-tentacles just leaves it as is. This modification does still leave the scenario of actually adding stages after the initial setup. To handle that case I have added a octopusci-reset-stage-locker cmd line tool.