Switched the octopusci-skel and octopusci-post-build-request commends to require octopusci/version rather than octopusci
I did this because when you require octopusci it attempts to load the /etc/octopusci/config.yml file and in the case of octopusci-skel which is supposed to generate that file it most likely doesn't exist. For octopusci-post-build-request the version info was the only part of octopusci that was necessary therefore I simply included that one thing instead of everything else.