Cleaned up octopusci-reset-redis code a bit.
This included replacing a require 'octopusci' with two requires 'octopusci/verison', and 'octopusci/job_store'. This makes the application a lot faster as loading the config using yaml seems to take a long time and that is not needed for the octopusci-reset-redis command so there is no real need to require 'octopusci' as an entire library. I also changed some triple quotes to a %Q evaluated string literal for standards.