Cleaned up octopusci-reset-stage-locker code a bit.
This involved adding a require 'octopusci/version' to the top and moving the require 'octopusci' lower into the actual conditional where reset is determined or not. This makes the script much more usable as it is super fast to get the help menu and see all the options. Beyond that it moves the little bit of waiting to the one place in the application where it makes sense, when the reset is happening.