Cleaned up the ux of the octopusci-skel cmd tool.
I reworked the app so that by default it does actually make any changes. You are now required to pass a switch to it, to make it actually perform the changes. I did this so that if a user is simply executing a command expecting to get a help screen so they know how to use it, it won't surprise them and actually make a bunch of changes to their system. I also updated the default config.yml content and now include a default job example which builds Octopusci itself.