Modified octopusci-post-build-request to build ocotpusci.
I modified the octopusci-post-build-request command line tool to make a request to bulid octopusci. I think I made all the necessary modifications though I haven't really tested this change yet due to some blockers. I made these changes so that this cmd line tool will eventually be able to be used as part of the install process and not only used for testing custom jobs but also for testing the default installation and by default trigger the default installed job for Octopusci.