Added the startings of octopusci-install
I got the basic structure of the cmd line tool setup and I got basic account creation working on Mac OS X. I have a lot more to do with this cmd line tool. But, it is a good start. The idea is that instead of running octopusci-skel after a gem install octopusci, you would simply run octopusci-install which would create the necessary user account, set up all the file structure necessary for octopusci's workspace and log storage as well as setup anything else that may be needed such as a thin server configuration, etc.