Added environment variable setting for specified daemon user.
I modified octopusci-tentacles so that it sets its HOME, SHELL, and USER environment variables based on the info it finds in /etc/passwd for the specified tentacles_user in the config.yml. I have also modified it to define the OCTOPUSCI environment variable so that when commands are executed by the jobs they can determine if they are being executed by Octopusci or not.