Hacked together a quick and dirty re-enqueue.
After using Octopusci for a long while now it is more obvious than ever
that the most important feature we are lacking at this point in time is
the ability to re-equeue a job. Therefore, I just hacked together an
ugly version of the feature this morning so that we could have
something to help out. It should be refactored and cleaned up I just
don't really have the time to do that right now.